+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Needed changes to make `ng build --prod` work
 - Changed repository structure
 - Make navbar position fixed
+- Reworked razorapi service for better handling of asynchronous events
 
 ## [0.1.0] - 2020-10-08
 ### Added